BUG: new 64 bit VR type - issues on Windows

+++ Source/DataStructureAndEncodingDefinition/gdcmVR.cxx
@@ -264,10 +264,10 @@ unsigned int VR::GetSizeof() const
   return size;
 }
 
-long VR::GetIndex(VRType vr)
+long long VR::GetIndex(VRType vr)
 {
   if( vr == VR::VL32 ) return 0;
-  long l;
+  long long l;
   assert( vr <= VR_END );
   switch(vr)
     {
@@ -291,9 +291,9 @@ long VR::GetIndex(VRType vr)
     break;
   default:
       {
-      long a = (long)vr;
+      long long a = (long long)vr;
       for (l = 0; a > 1; ++l)
-        a >>= 1;
+        a >>= 1LL;
       l++;
       }
     }
@@ -375,7 +375,7 @@ VR::VRType VR::GetVRTypeFromFile(const char *vr)
     // Use lazy evaluation instead of strncmp
     if (ref[0] == vr[0] && ref[1] == vr[1] )
       {
-      r = (VR::VRType)(1L << (i-1));
+      r = (VR::VRType)(1LL << (i-1));
       break;
       }
     }
@@ -433,7 +433,7 @@ VR::VRType VR::GetVRType(const char *vr)
         break;
       default:
         assert( vr[2] == 0 );
-        r = (VR::VRType)(1L << (i-1));
+        r = (VR::VRType)(1LL << (i-1));
         }
       break; // found one value, we can exit the for loop
       }
